+++ feature-newsfeed/src/main/java/ru/sviridov/vkclient/feature_newsfeed/data/db/entity/NewsItemEntity.kt
@@ -8,7 +8,7 @@ class NewsItemEntity(
     @PrimaryKey val postId: Int,
     val sourceId: Int,
     var sourceTitle: String,
-    val postedAt: Int,
+    val postedAt: Long,
     var sourceAvatar: String,
     var imageUrl: String?,
     var textContent: String?,
